+++ CHANGELOG.md
@@ -2,6 +2,7 @@ UNRELEASED
 ----------
 
 * [ [#1108](https://github.com/digitalfabrik/integreat-cms/issues/1108) ] Support SVG images in PDF export
+* [ [#1284](https://github.com/digitalfabrik/integreat-cms/issues/1284) ] Inherit status of new translations from source language on XLIFF import
 
 
 2022.3.3